feat: 增加 SafeMSet/SafeMGet 函数

+++ cmd_bitmap_test.go
@@ -243,7 +243,9 @@ func doTestUnitClean(ctx context.Context, c Cmdable, keys []string) {
 	if len(keys) > 0 {
 		So(c.Del(ctx, keys...).Err(), ShouldBeNil)
 	}
-	c.FlushAll(context.Background())
+	if !c.Options().GetDevelopment() {
+		c.FlushAll(context.Background())
+	}
 }
 
 type TestUnitName interface {
@@ -269,16 +271,27 @@ func bitMapTestUnits() []TestUnit {
 	}
 }
 
-func doTestUnits(t *testing.T, unitsFunc func() []TestUnit) {
-	c := MustNewClient(NewConf(WithDevelopment(false)))
-	c.FlushAll(context.Background())
+func _doTestUnits(t *testing.T, c Cmdable, unitsFunc func() []TestUnit) {
 	t.Cleanup(func() {
 		_ = c.Close()
 	})
+	if !c.Options().GetDevelopment() {
+		c.FlushAll(context.Background())
+	}
 	var ctx = context.Background()
 	for _, v := range unitsFunc() {
 		Convey(v.Name.String(), t, func() { doTestUnitClean(ctx, c, v.Func(ctx, c)) })
 	}
+}
+
+func doTestUnits(t *testing.T, unitsFunc func() []TestUnit) {
+	c := MustNewClient(NewConf(WithDevelopment(false)))
+	_doTestUnits(t, c, unitsFunc)
+}
+
+func doClusterTestUnits(t *testing.T, unitsFunc func() []TestUnit) {
+	c := MustNewClient(NewConf(WithDevelopment(true)))
+	_doTestUnits(t, c, unitsFunc)
 }
 
 func TestClient_BitMap(t *testing.T) { doTestUnits(t, bitMapTestUnits) }
